requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.2 # для PostgreSQL; для SQLite не обязателен
python-docx>=1.1.0
Pillow>=10.3.0
djangorestframework>=3.15.2
python-dotenv>=1.0.1